requirements.txt
google-generativeai
colorama
chromadb
selenium
pillow
pymetasploit3
msgpack
streamlit
pandas
watchdog
